Add polling watcher, Redis lock stub, and API updates
- Introduced MongoDocumentWatcherPolling for polling-based change detection as an alternative to change streams.
- Split watcher into MongoDocumentWatcherStream (streams) and MongoDocumentWatcherPolling (polling).
- Added AddWatchPolling() and AddWatchStream() DI extensions.
- Added WatchPollInterval to MongoObjectOptions.
- Updated README with polling watcher docs and roadmap.
- Added RedisDistributedLockManager stub and AddRedisLockManager extension in new MongoObject.RedisDistributedLock project.
- Enhanced IDistributedLockManager interface for new lock operations.
- Changed Version property in IMetadataBase and generated code from int? to long?.
- Updated IDocumentFile and IMongoConnection interfaces.
- Required explicit IMongoClient registration in examples.
- Cleaned up project files and solution references.
- Added PolingSetup example and improved cache error handling.

+++ MongoObject.Core/Interfaces/IMetadataBase.cs
@@ -46,6 +46,6 @@ namespace MongoObject.Core.Interfaces
     {
         public DateTime? CreatedAt { get; set; }
         public DateTime? LastModifiedAt { get; set; }
-        public int? Version { get; set; }
+        public long? Version { get; set; }
     }
 }
